Fix travis ci error
