--- docker/examples/general-v3.go
@@ -1,23 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"github.com/sendgrid/sendgrid-go"
-	"log"
-)
-
-func main() {
-	apiKey := "JUST_A_TEST_KEY"
-	host := "http://localhost:4010"
-	request := sendgrid.GetRequest(apiKey, "/v3/api_keys", host)
-	request.Method = "GET"
-
-	response, err := sendgrid.API(request)
-	if err != nil {
-		log.Println(err)
-	} else {
-		fmt.Println(response.StatusCode)
-		fmt.Println(response.Body)
-		fmt.Println(response.Headers)
-	}
-}
